fix(theme): remove wario theme from ds lib
