+++ packages/react/src/index.ts
@@ -95,7 +95,7 @@ export { Pagination } from './components/pagination/pagination';
 
 // Themes
 export {
-    equisoftTheme, defaultTheme, warioTheme, buildTheme,
+    equisoftTheme, defaultTheme, buildTheme,
 } from './themes';
 export { injectMainCss } from './styles';
 export { ResolvedTheme, ThemeCustomization, ThemeCustomization as Theme } from './themes/theme';
