+++ packages/react/src/themes/index.ts
@@ -1,5 +1,4 @@
 export * from './equisoft';
 export * from './default-theme';
-export * from './wario-theme';
 export * from './tokens';
 export * from './build-theme';
